+++ CHANGELOG.md
@@ -24,7 +24,8 @@ Improvements:
 * Core: updated default favicon.
 * Pages: updated pages getNavigation; the 'includeChildren' parameter was useless.
 * Core: refactored javascript ajax-calls.
-
+* Installer: refactored installation of dashboard widgets.
+* Analytics: dashboard widgets are now added for all users upon installation.
 
 2.6.13 (2011-10-18)
 --